refactor: adopt `v0` folder for code generation at this stage
