+++ buf.gen.yaml
@@ -6,15 +6,15 @@
 version: v2
 plugins:
   - remote: buf.build/protocolbuffers/go:v1.31.0
-    out: gen
+    out: v0
     opt:
       - paths=source_relative
   - remote: buf.build/grpc/go:v1.3.0
-    out: gen
+    out: v0
     opt:
       - paths=source_relative
   - remote: buf.build/grpc-ecosystem/gateway:v2.26.1
-    out: gen
+    out: v0
     opt:
       - paths=source_relative
       - generate_unbound_methods=true
