+++ buf.yaml
@@ -5,9 +5,7 @@
 # https://buf.build/docs/configuration/v2/buf-yaml
 version: v2
 modules:
-  - path: proto
-    excludes:
-      - proto/google
+  - path: proto/v0
 lint:
   use:
     - STANDARD
